fix: Remove @mui/icons-material
It's a great library but it increases our CI build time, stores an extra 128 MB of deps on disk, and it's unlikely we will ever use that many icons in this repo. Adding the few we do care about isn't that much work.

Minor rework to the Third Party page to add in line css that was removed by this package removal.